only build stable version; multiple pypy versions
This fixes the pypi version parsing to only look for stable versions
(recently the update script has been failing to build scipy because it
picks up the 1.0.0b1 and rc1 versions, but a build doesn't actually
build those).

It also adds support for building wheels for multiple pypy versions, and
adds a 5.9.0 build.